+++ rust/ommx/src/convert/parametric_instance.rs
@@ -1,17 +1,8 @@
 use crate::{
-    random::{
-        arbitrary_constraints, arbitrary_decision_variables, arbitrary_parameters,
-        FunctionParameters,
-    },
-    v1::{
-        decision_variable::Kind,
-        instance::{Description, Sense},
-        Function, Instance, Parameters, ParametricInstance, State,
-    },
+    v1::{Function, Instance, Parameters, ParametricInstance, State},
     Evaluate,
 };
 use anyhow::{bail, Result};
-use proptest::prelude::*;
 use std::{borrow::Cow, collections::BTreeSet};
 
 impl From<Instance> for ParametricInstance {
@@ -165,104 +156,11 @@ impl ParametricInstance {
     }
 }
 
-impl Arbitrary for ParametricInstance {
-    type Parameters = (usize, usize, u32, u64);
-    type Strategy = BoxedStrategy<Self>;
-
-    fn arbitrary_with(
-        (num_constraints, num_terms, max_degree, max_id): Self::Parameters,
-    ) -> Self::Strategy {
-        (
-            proptest::option::of(Function::arbitrary_with(FunctionParameters {
-                num_terms,
-                max_degree,
-                max_id,
-            })),
-            arbitrary_constraints(
-                num_constraints,
-                FunctionParameters {
-                    num_terms,
-                    max_degree,
-                    max_id,
-                },
-            ),
-        )
-            .prop_flat_map(|(objective, constraints)| {
-                let mut used_ids = objective
-                    .as_ref()
-                    .map(|f| f.used_decision_variable_ids())
-                    .unwrap_or_default();
-                for c in &constraints {
-                    used_ids.extend(c.function().used_decision_variable_ids());
-                }
-
-                (
-                    Just(objective),
-                    Just(constraints),
-                    arbitrary_split(used_ids),
-                )
-                    .prop_flat_map(
-                        |(objective, constraints, (decision_variable_ids, parameter_ids))| {
-                            (
-                                Just(objective),
-                                Just(constraints),
-                                arbitrary_decision_variables(
-                                    decision_variable_ids,
-                                    Kind::arbitrary(),
-                                ),
-                                arbitrary_parameters(parameter_ids),
-                                Option::<Description>::arbitrary(),
-                                Sense::arbitrary(),
-                            )
-                                .prop_map(
-                                    |(
-                                        objective,
-                                        constraints,
-                                        decision_variables,
-                                        parameters,
-                                        description,
-                                        sense,
-                                    )| {
-                                        ParametricInstance {
-                                            objective,
-                                            constraints,
-                                            decision_variables,
-                                            description,
-                                            sense: sense as i32,
-                                            parameters,
-                                            ..Default::default()
-                                        }
-                                    },
-                                )
-                        },
-                    )
-            })
-            .boxed()
-    }
-}
-
-fn arbitrary_split(ids: BTreeSet<u64>) -> BoxedStrategy<(BTreeSet<u64>, BTreeSet<u64>)> {
-    let flips = proptest::collection::vec(bool::arbitrary(), ids.len());
-    flips
-        .prop_map(move |flips| {
-            let mut used_ids = BTreeSet::new();
-            let mut defined_ids = BTreeSet::new();
-            for (flip, id) in flips.into_iter().zip(ids.iter()) {
-                if flip {
-                    used_ids.insert(*id);
-                } else {
-                    defined_ids.insert(*id);
-                }
-            }
-            (used_ids, defined_ids)
-        })
-        .boxed()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use approx::abs_diff_eq;
+    use proptest::prelude::*;
 
     proptest! {
         #[test]
